Refactor round 1
