+++ src/os_open_usrn/create_motherduck_table.py
