+++ src/street_manager/extract_load_data.py
@@ -1,46 +1,11 @@
 import json
-import requests
 import pandas as pd
+
+from flatten_street_manager_data import flatten_json
 
 from stream_unzip import stream_unzip
 from loguru import logger
 from tqdm import tqdm
-
-
-def fetch_data(dl_url):
-    """
-    Stream data from DfT website for the Street Manager data you want
-
-    Args: 
-        takes the url for the street manager data for exmaple:
-        "https://opendata.manage-roadworks.service.gov.uk/permit/2024/03.zip"
-    
-    It should return chunks of the data to be processed further.
-    """
-    with requests.get(dl_url, stream=True, timeout=30) as chunk:
-        yield from chunk.iter_content(chunk_size=1048576)
-
-
-def flatten_json(json_data) -> dict:
-    """
-    Street manager archived open data comes in nested json files
-    This function flattens the structure
-    
-    Args:
-        json_data to flatten
-    
-    """
-    flattened_data = {}
-
-    def flatten(data, prefix=''):
-        if isinstance(data, dict):
-            for key in data:
-                flatten(data[key], f'{prefix}{key}.')
-        else:
-            flattened_data[prefix[:-1]] = data
-
-    flatten(json_data)
-    return flattened_data
 
 
 def quick_col_rename(df) -> pd.DataFrame:
@@ -52,44 +17,6 @@ def quick_col_rename(df) -> pd.DataFrame:
     This will remove "object data." from the column names
     """
     df.columns = [col.replace("object_data.", "") if "object_data." in col else col for col in df.columns]
-    return df
-
-
-def check_data_schema(zipped_chunks):
-    """
-    Reads 50 JSON files from zipped chunks and returns a Pandas DataFrame.
-    
-    This is so you can assess the data structure and validate part of it against the Pydantic model. 
-    
-    Args:
-        zipped_chunks: Iterable of zipped chunks containing JSON files.
-    """
-    max_files = 50
-    file_count = 0
-    data_list = []
-
-    for file, size, unzipped_chunks in tqdm(stream_unzip(zipped_chunks)):
-        if file_count >= max_files:
-            break
-
-        if isinstance(file, bytes):
-            file = file.decode('utf-8')
-
-        try:
-            # Decode bytes to string and load into JSON
-            bytes_obj = b''.join(unzipped_chunks)
-            json_data = json.loads(bytes_obj.decode('utf-8'))
-            # Flatten the JSON data if necessary and add to list
-            flattened_data = flatten_json(json_data)
-            data_list.append(flattened_data)
-            file_count += 1
-
-        except Exception as e:
-            print(f"Error processing {file}: {e}")
-            raise
-
-    # Create DataFrame from the collected data
-    df = pd.DataFrame(data_list)
     return df
