+++ src/street_manager/motherduck_create_table.py
