+++ src/street_manager/schema_name_trigger.py
