+++ src/street_manager/stream_zipped_data.py
@@ -1,0 +1,21 @@
+import json
+import requests
+import pandas as pd
+
+from stream_unzip import stream_unzip
+from loguru import logger
+from tqdm import tqdm
+
+
+def fetch_data(dl_url):
+    """
+    Stream data from DfT website for the Street Manager data you want
+
+    Args: 
+        takes the url for the street manager data for exmaple:
+        "https://opendata.manage-roadworks.service.gov.uk/permit/2024/03.zip"
+    
+    It should return chunks of the data to be processed further.
+    """
+    with requests.get(dl_url, stream=True, timeout=30) as chunk:
+        yield from chunk.iter_content(chunk_size=1048576)
